Build for Android/Windows 64-bit
Fix build error for Android 64-bit

Fix build again

Update CMakeLists

Update CMakeLists

Fix annoying warning

Update CMakeLists

Update CMakeLists

Update CMakeLists

Update CMakeLists

Fix memory leak when reading .nav files

Fix loading filesystem_stdio on Android

Disable warning

RemovePlayerItemEx: Reduce ammo supply for grenades and c4

RemovePlayerItemEx: Do not remove the grenade weapons until it runs out of ammo

Fix windows 32-bit

Simplify CMakeLists

Fix android oopsie

Fix

I hope final build fix

OK this time that's it

+++ regamedll/game_shared/bot/nav_file.cpp
@@ -158,8 +158,8 @@ void CNavArea::Save(int fd, unsigned int version)
 	for (int d = 0; d < NUM_DIRECTIONS; d++)
 	{
 		// save number of connections for this direction
-		size_t count = m_connect[d].size();
-		_write(fd, &count, sizeof(size_t));
+		unsigned int count = m_connect[d].size();
+		_write(fd, &count, sizeof(unsigned int));
 
 		for (auto &connect : m_connect[d]) {
 			_write(fd, &connect.area->m_id, sizeof(unsigned int));
